Adds browser-cache-circumvention to rx-jupyter.
This is one of several ways we might fix #3850.

In this case we assume that Jupyter server will not change its ways
regarding lack of Cache-control headers on APIs. So we augment the
querystring of GET and HEAD requests w/ a timestamp to ensure the
request makes it to the server. This is how Jupyter notebook does it.

I mechanically went through https://github.com/jupyter/notebook/tree/master/notebook/static/services
and found usages of jquery's ajax method w/ its `cache` option set to
false. Notably this was occurring on several PUT/POST/DELETE calls;
jquery docs say that doesn't make sense except for IE8, and their
code actually NOOPs these cases, so I've ommitted them.

I altered createAJAXSettings to behave much like jquery's ajax().
This creates a familiar API (for many), and centralizes the cache-
busting logic. It's notable however that in many cases we are now
assembling a URL string, only to pass it into createAJAXSettings
where it is parsed, the cache-busting query section added, before
a full string is finally rebuilt.