Fix regression in tab completion
`switchMap` sends 4 arguments on each item passed through. We only want the first and don't care about the rest. Without this, the `of` gets all the arguments which ends up turning each of the arguments into new items in the stream, causing the whole stream to fail later.

We probably should disable `no-unnecessary-callback-wrapper` for our codebase because
it makes an assumption about the number of arguments.

This came from:

https://github.com/nteract/nteract/pull/4083/files#r249873790

Which, applied as part of our tsconfig made sense. We have to be careful about this rule though, and I think it's an area where a human reviewer around unnecessary callback wrappers would be better.