Pin react-table and vendorize stylesheet deps
This is a quick and dirty copy/paste of the react-table and react-table-hoc-fixed-columns stylesheet dependencies.

In the future, instead of hard pinning these react-table dependencies, we could set up a script to produce the vendorized css files to account for any css updates (see @nteract/styled-blueprintjsx as an example). I thought I would keep it simple for now and iterate since this would have no visible difference to the user.

Fixes #4306