Change plotly dependency to official distribution
The version of plotly referenced in the plotly-transform package
lagged the official version. Now that plotly publishes a version
of its minified js to npm it made sense to change the dependency
to the official plotly package

This change required mocking out a bit more of the window object.
This is a well known Jest/JSDOM problem