Cleanup the action definitions in hosts.ts
This PR promotes the clean concise action definition style
to avoid code repetition as proposed in #4561.

The new pattern is only applied to action definitions in `hosts.ts`.

Question to reviewers:
This file contains four bookstore actions and one host action which is not consistent. Would it make sense to split it?