Fix desktop integration in AppImage package
Remove the hardcoded reference to /opt/nteract in the configuration for electron-builder. This hardcoded
reference broke the desktop integration for the AppImage package, which is not installed to /opt/nteract.

The "Path"-key in the .desktop file, which controls the working directory where the app should be started,
is not necessary because nteract does not need to be run from a specific working directory.